refactor: rename the jwk helper module
